Fix run parts of install only once

+++ install.sh
@@ -36,10 +36,15 @@ done
 source ${HOME}/.bashrc
 
 
-# Update
-sudo apt update && sudo apt upgrade -y && sudo apt full-upgrade -y && sudo apt autoremove -y
+if [ ! -f "${PWD}/.install_did_run" ]; then
+  touch "${PWD}/.install_did_run"
+  # Update
+  sudo apt update && sudo apt upgrade -y && sudo apt full-upgrade -y && sudo apt autoremove -y
 
-sudo apt install -y curl git htop vim mosh wget tmux clang-format
+  sudo apt install -y curl git htop vim mosh wget tmux clang-format
+
+  echo -e "\n\n192.168.47.62   vagrant\n" | sudo tee -a /etc/hosts
+fi
 
 
 echo "
